msm: camera: isp: For testgen disable camif on frame boundary
When testgen is used then we cannot gate the input testgen on a frame
boundary and then stop camif. In this case if camif is stopped
immediately then it causes hardware failures on next camif start.
To avoid these errors from hardware always stop camif on frame
boundary when testgen is used.

CRs-Fixed: 987962
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
Change-Id: I288d5883742bf1d55eda8acdfb687c168551c9c2
Signed-off-by: Lau <laststandrighthere@gmail.com>