pstore: Protect unlink with read_mutex
When update_ms is set, pstore_get_records() will be called when there's
a new entry.  But unlink can be called at the same time and might
contend with the open-read-close loop.  Depending on the implementation
of platform driver, it may be safe or not.  But I think it'd be better
to protect those race in the first place.

Cc: Stefan Hajnoczi <stefanha@redhat.com>
Signed-off-by: Namhyung Kim <namhyung@kernel.org>
Change-Id: I083a6738d194b2fbe7d74abefa889ebe9c01127b
Signed-off-by: Kees Cook <keescook@chromium.org>
(cherry picked from commit e9e360b08a44098ec6f31de8e5a29a3ffaada828)
Signed-off-by: TogoFire <togofire@mailfence.com>