ida: Use exceptional entries for small IDAs
We can use the root entry as a bitmap and save allocating a 128 byte
bitmap for an IDA that contains only a few entries (30 on a 32-bit
machine, 62 on a 64-bit machine).  This costs about 300 bytes of kernel
text on x86-64, so as long as 3 IDAs fall into this category, this
is a net win for memory consumption.

Thanks to Rasmus Villemoes for his work documenting the problem and
collecting statistics on IDAs.

Signed-off-by: Matthew Wilcox <mawilcox@microsoft.com>
Change-Id: I3632ecbbeacad8ac1340e110dc667d91295c514a
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>
(cherry picked from commit eeed06588aecbc0101060503ca97072b36d7877e)
Signed-off-by: TogoFire <togofire@mailfence.com>