crypto: skcipher - introduce walksize attribute for SIMD algos
In some cases, SIMD algorithms can only perform optimally when
allowed to operate on multiple input blocks in parallel. This is
especially true for bit slicing algorithms, which typically take
the same amount of time processing a single block or 8 blocks in
parallel. However, other SIMD algorithms may benefit as well from
bigger strides.

So add a walksize attribute to the skcipher algorithm definition, and
wire it up to the skcipher walk API. To avoid confusion between the
skcipher and AEAD attributes, rename the skcipher_walk chunksize
attribute to 'stride', and set it from the walksize (in the skcipher
case) or from the chunksize (in the AEAD case).

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Change-Id: Ibe16c7dc55ca7c3323cc3a3b761c768eb9b7bb88
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
(cherry picked from commit c821f6ab2e47946f35ee2f30781c5185e5d07f65)
Signed-off-by: TogoFire <togofire@mailfence.com>