radix_tree_iter_resume: Fix out of bounds error
The address sanitizer occasionally finds an out of bounds error while
running the test-suite.  It turned out to be a read of the pointer
immediately next to the tree root, but this out of bounds error could
have occurred elsewhere.  This happens because radix_tree_iter_resume()
dereferences 'slot' before checking whether we've come to the end of
the chunk.  We can just delete this line; the value was never used.

Signed-off-by: Matthew Wilcox <mawilcox@microsoft.com>
Change-Id: I456605f5c56694aed9c6d0f06144307f484e9f12
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>
(cherry picked from commit 97bea25df36b518a7aa6858af12a89948e22ccec)
Signed-off-by: TogoFire <togofire@mailfence.com>