mbcache: make mbcache naming more generic
Make names more generic so that mbcache usage is not limited to
block sharing. In a subsequent patch in the series
("ext4: xattr inode deduplication"), we start using the mbcache code
for sharing xattr inodes. With that patch, old mb_cache_entry.e_block
field could be holding either a block number or an inode number.

Signed-off-by: Tahsin Erdogan <tahsin@google.com>
Change-Id: Ie432620bf5376d4c20a1b8c533f667eadabb20d4
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: TogoFire <togofire@mailfence.com>