sched/fair: Fix issue where frequency update not skipped
This patch fixes one of the infrequent conditions in
commit 54b6bae ("sched/fair: Skip frequency updates if CPU about to idle")
where we could have skipped a frequency update. The fix is to use the
correct flag which skips freq updates.

Note that this is a rare issue (can show up only during CFS throttling)
and even then we just do an additional frequency update which we were
doing anyway before the above patch.

Bug: 64689959

Signed-off-by: Joel Fernandes <joelaf@google.com>
Change-Id: Ic1e6926d6f57de1784bb3118f9336ebef6c772ea
Signed-off-by: Lau <laststandrighthere@gmail.com>