ANDROID: init: ensure initcall ordering with LTO
With LTO, the compiler doesn't necessarily obey link order for
initcalls, and the initcall variable needs to be globally unique
to avoid naming collisions.

In order to preserve the correct order, we add each variable
into its own section and generate a linker script (in
scripts/link-vmlinux.sh) to ensure the order remains correct.  We
also add a __COUNTER__ prefix to the name, so we can retain the
order of initcalls within each compilation unit, and __LINE__ to
make the names more unique.

Bug: 145210207
Signed-off-by: Sami Tolvanen <samitolvanen@google.com>
Change-Id: I0060ce8f6c27e0df5993c8f107c9839297cc3883
Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>
Signed-off-by: CloudedQuartz <ravenklawasd@gmail.com>
Signed-off-by: GeoPD <geoemmanuelpd2001@gmail.com>