thermal: msm_lmh_dcvs: Never initialize LMH TZ more than once
The LMH DCVS initialization consists in enabling the LMH thermal,
outer loop, reliability and BCL algorithms on a per-cluster
basis, but the command that kickstarts it is "PROFILE_CHANGE"
with parameter '1' sent to the TZ LMH Service.

Sending it more than one time is rendundant, but the critical
issue is that - if we send it more than once - the CPUs will
deadlock!

Change-Id: 7f0dbf13b062e099a8f0dfe5dd0be7a54e41691f
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: TogoFire <togofire@mailfence.com>