sched: walt: Optimize task_util()
task_util() for WALT is currently defined as

p->ravg.demand / (sched_ravg_window >> SCHED_CAPACITY_SHIFT);

This math is required to scale the task demand to 1024 scale.
task_util() is used many times in task placement. So the calls
to this can be optimized by caching the scaled value when task
demand is calculated.

Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Change-Id: I999cb2948eef7044b5f0f001dd2c968ec08cf3e3
Signed-off-by: TogoFire <italomellopereira@gmail.com>