[BACKPORT]: drm/msm/sde: add sysfs node in dpu driver to provide fps
Add sysfs node on crtc to get fps value at runtime.
The algorithm will return last committed frame count in
the configured periodicity timeline. The calculation is
time based where in dpu driver counts number of frames
for the given periodicity value, otherwise it counts the
number of frames transferred for last one second.

Below are the commands to give in shell:

Command to set periodicity:
	echo value > /sys/class/drm/sde-crtc-*/fps_periodicity_ms

Command to get periodicity:
	cat /sys/class/drm/sde-crtc-*/fps_periodicity_ms

Command to get the measured fps:
	cat /sys/class/drm/sde-crtc-*/measured_fps

Replace * with 0, 1 or 2.

Signed-off-by: Shubhashree Dhar <dhar@codeaurora.org>
Change-Id: I2b631e5c43d01cbc2cff625ffea3ab24480c2672
Signed-off-by: Raphiel Rollerscaperers <rapherion@raphielgang.org>
Signed-off-by: PainKiller3 <ninadpatil100@gmail.com>