power: qpnp-fg-gen3: fix soc not keeping full after charging overnight
The issue happened when the temperature was cool after charge_done.
(chip->health was not POWER_SUPPLY_HEALTH_GOOD). In this condition,
chip->charge_full was not set, and then battery level dropped.

Set chip->charge_full to true if chip->health is
POWER_SUPPLY_HEALTH_GOOD or POWER_SUPPLY_HEALTH_COOL
or POWER_SUPPLY_HEALTH_WARM to fix it.

Bug: 122704179
Signed-off-by: Jack Wu <wjack@google.com>
Change-Id: I11b5a3e565479c7b09e5f8940c57467d3cf26d1d
Signed-off-by: Lau <laststandrighthere@gmail.com>