msm: kgsl: Skip state change in idle check, if requested state is NONE
After scheduling the idle work for transitioning to either NAP or SLUMBER
it's possible that requested state can change to NONE if any new workload
comes before kgsl_idle_check is executed or it gets the device mutex.
If this happens, current code calls pwrctrl_change_state which fails as
this is not a legal transition. To prevent this, skip calling
pwrctrl_change_state if requested state is NONE.

Signed-off-by: Deepak Kumar <dkumar@codeaurora.org>
Change-Id: I8bddb668c21480f60cc428c36b5dc48ae6c03c74
Signed-off-by: Raphiel Rollerscaperers <rapherion@raphielgang.org>
Signed-off-by: palaych <palaychm@yandex.ru>