cpu_input_boost: Track input event times and expose timeout helper
Sometimes we may want to limit certain boost kicks to only take effect
within a certain timeout after receiving the last input event to save
power. For example, we may want to boost when an intensive process
forks to improve fluidity, but boosting when the fork was not initiated
by input is wasteful.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Change-Id: If93de1a715e906f9b6f1ee4a540e7a76f8b9e2a9
Signed-off-by: Lau <laststandrighthere@gmail.com>