cpu_input_boost: Allow configuration of CPU base frequency
On some CPUs, it can be more efficient to use a higher base frequency
after all boosts have expired. It is actually possible for battery life to
improve when using this elevated base frequency because there are less
CPU frequency transitions (which are relatively expensive) and less
"ping-pong" effects where the CPU constantly switches between
frequencies in the 300000 KHz - 576000 KHz range. This has been observed
to occur during video playback in some cases.

This can also be used to bump the performance baseline for intensive
tasks such as gaming.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Change-Id: I6de193afdfe2201a8ae34e5eaa419f6bb5c9a6c7
Signed-off-by: laststandrighthere <laststandrighthere@gmail.com>
Signed-off-by: Lau <laststandrighthere@gmail.com>