msm: kgsl: Don't allocate memory dynamically for temp command buffers
The temporary command buffer in _set_pagetable_gpu is only the size of a
single page, and _set_pagetable_gpu is never executed concurrently. It
is therefore easy to replace the dynamic command buffer allocation with
a static one to improve performance by avoiding the latency of dynamic
memory allocation.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: Ifdc45fc543f34e3e0916f4d041600fb95da83a10
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Lau <laststandrighthere@gmail.com>