f2fs: Fix absolute jiffy duration in write uncongestion timeout
There's too much variation here when the duration is hard-coded as a
number of jiffies:
  100 Hz  => 10 ms
  250 Hz  => 8 ms
  300 Hz  => ~6.7 ms
  1000 Hz => 2 ms

The bigger concern is that on 1000 Hz, the delay becomes too short to
be meaningful and may cause issues.

Fix the issue by using a common duration target in milliseconds (6 ms)
combined with the msecs_to_jiffies() helper, which will calculate the
number of jiffies closest to the above duration on any given timer
frequency.

The new times are as follows:
  100 Hz  => 1 jiffy   (effective: 10 ms)
  250 Hz  => 2 jiffies (effective: 8 ms)
  300 Hz  => 2 jiffies (effective: 6 ms)
  1000 Hz => 6 jiffies (effective: 6 ms)

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Change-Id: Ie34312c221ac5a0dcc0448b08412a23dfb88b5d0
Signed-off-by: TogoFire <italomellopereira@gmail.com>