UPSTREAM: wireguard: receive: remove dead code from default packet type case
The situation in which we wind up hitting the default case here
indicates a major bug in earlier parsing code. It is not a usual thing
that should ever happen, which means a "friendly" message for it doesn't
make sense. Rather, replace this with a WARN_ON, just like we do earlier
in the file for a similar situation, so that somebody sends us a bug
report and we can fix it.

Reported-by: Fabian Freyer <fabianfreyer@radicallyopensecurity.com>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: I07ff93a0233d3a3d93bd9ff3babd38785d142584
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 2b8765c52db24c0fbcc81bac9b5e8390f2c7d3c8)
Bug: 152722841
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: I07ff93a0233d3a3d93bd9ff3babd38785d142584
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
(cherry picked from commit 9c059358938ef01c4676bde3a475726b51e110d5)
Signed-off-by: TogoFire <togofire@mailfence.com>