pinctrl: msm: Remove explicit barriers from mmio ops where unneeded
For the vast majority of mmio operations in this driver, explicit memory
barriers aren't needed either because a data dependency between a read
and write already exists, or because of the presence of the spin locks
which execute a full memory barrier.

Removing all the unneeded explicit barriers considerably reduces
overhead for pinctrl operations, which in turn benefits things like i2c.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I32520db9b83c9667f5652eae08996dcdb3146ae8
[Lau: fix conflicts for 4.9]
Signed-off-by: Lau <laststandrighthere@gmail.com>
[TogoFire: add what was missing]
Signed-off-by: TogoFire <togofire@mailfence.com>