msm: kgsl: Avoid race of fault handler and recovery
There is a possibility of cx gdsc off timeout when stall on fault
is enabled. smmu fault handler is blocked on device mutex to keep
cx vote and dispatcher trying to turn it off for recovery. To avoid
this race condition, return from fault handler when there is contention
for device mutex.

Signed-off-by: gkiranku <gkiranku@codeaurora.org>
Change-Id: I45ecf9e20f33d51828a185c99dffe4706fe2a419
Signed-off-by: TogoFire <togofire@mailfence.com>