drivers: mmc: skip all power management when SUSPEND_SKIP_SYNC is set
since it's completely broken when said config is set and even when all the functions which "waits till sync is done" (specially the workqueue ones) are converted to non-wait ones, it just keeps trying to suspend and resume everytime and fails to do so in the end, wasting power and precious suspend cycles, so just skip the whole suspend and resume functions during this state.

deep sleep drain remains unchanged, at min 20ma on average.

nuking pm props for all the files in the folder (we only use a certain portion of it but it's harmless when compiled with stock config as SKIP_SYNC is not set by default)


















-e 
Change-Id: I7a16260cd5e26be1e50435f03494372c3bc7d6ef