socket: remove bogus __be32 annotation
The endpoint->src_if4 has nothing to do with fixed-endian numbers; remove
the bogus annotation.

This was introduced in
https://git.zx2c4.com/wireguard-monolithic-historical/commit?id=14e7d0a499a676ec55176c0de2f9fcbd34074a82
in the historical WireGuard repo because the old code used to
zero-initialize multiple members as follows:

    endpoint->src4.s_addr = endpoint->src_if4 = fl.saddr = 0;

Because fl.saddr is fixed-endian and an assignment returns a value with the
type of its left operand, this meant that sparse detected an assignment
between values of different endianness.

Since then, this assignment was already split up into separate statements;
just the cast survived.

Signed-off-by: Jann Horn <jannh@google.com>
Change-Id: I471227ff19c4aed40fdc68f00c47aeb2923b4c87
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
(cherry picked from commit 582d2e75efcfd4c1dfab2963539c029c43abe066)
Signed-off-by: TogoFire <togofire@mailfence.com>