prlmk: Handle a critical lowmemory situation
Now that we've set the appropriate notifiers in is_low_mem(),
we use that to tell us when the memory situation is critically
low, and accordingly call pagefault_out_of_memory(), which will
kill tasks that consumes the most amount of memory (i.e, memory hogs).

pagefault_out_of_memory() is defined in mm/oom_kill.c

This is a temporary fix until a better alternative can be found.

Signed-off-by: darkhz <kmachanwenw@gmail.com>
Change-Id: Id6678da5cf0ba3a6b6c8c78b9dcb8c04fa59f78b