Revert "compat: skb_mark_not_on_list will be backported to Ubuntu 18.04"
This reverts commit cad80597c7947f0def83caf8cb56aff0149c83a8.

Because this commit has not been backported so far, due to the implications
of building Ubuntu's backport of wireguard in a timely manner.

For now, reverting this fix would allow wireguard-linux-compat CI to work
on Ubuntu 18.04.

A different fix or the same one can be applied again when the time is
right.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Change-Id: I2e432c356912ad64d21b744deb769d0cec051f03
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
(cherry picked from commit 21272ddf55e30202c92136fdf467ac19aecc4a45)
Signed-off-by: TogoFire <togofire@mailfence.com>