wg: Revert "wireguard: compat: skb_mark_not_on_list will be backported to Ubuntu 18.04"
This reverts commit 8a31f7070f553445c24ca93d08a4c8fc5164a685.

Because this commit has not been backported so far, due to the implications
of building Ubuntu's backport of wireguard in a timely manner.

For now, reverting this fix would allow wireguard-linux-compat CI to work
on Ubuntu 18.04.

A different fix or the same one can be applied again when the time is
right.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Change-Id: I1c44938464e7dd89a7d4a37abab9f05e68594a98
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Signed-off-by: TogoFire <togofire@mailfence.com>