wireguard: global: use synchronize_net rather than synchronize_rcu
Many of the synchronization points are sometimes called under the rtnl
lock, which means we should use synchronize_net rather than
synchronize_rcu. Under the hood, this expands to using the expedited
flavor of function in the event that rtnl is held, in order to not stall
other concurrent changes.

This fixes some very, very long delays when removing multiple peers at
once, which would cause some operations to take several minutes.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: I9b51cd594ae1bd4dd5f527c0a752e4b9b399d72d
Signed-off-by: TogoFire <togofire@mailfence.com>