wireguard: main: rename 'mod_init' & 'mod_exit' functions to be module-specific
Rename module_init & module_exit functions that are named
"mod_init" and "mod_exit" so that they are unique in both the
System.map file and in initcall_debug output instead of showing
up as almost anonymous "mod_init".

This is helpful for debugging and in determining how long certain
module_init calls take to execute.

Signed-off-by: Randy Dunlap <rdunlap@infradead.org>
Change-Id: I31eaeb8ad950c038a51a8346150bb181bb102cf4
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Signed-off-by: TogoFire <togofire@mailfence.com>