UPSTREAM: HID: playstation: add microphone mute support for DualSense
The DualSense controller has a built-in microphone exposed as an
audio device over USB (or HID using Bluetooth). A dedicated
button on the controller handles mute, but software has to configure
the device to mute the audio stream.
This patch captures the mute button and schedules an output report
to mute/unmute the audio stream as well as toggle the mute LED.

Bug: 167947264
Signed-off-by: Roderick Colenbrander <roderick.colenbrander@sony.com>
Change-Id: I11f964d5a277542b307f4fb88afbba9a0966ff2c
Reviewed-by: Barnabás Pőcze <pobrn@protonmail.com>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Farid Chahla <farid.chahla@sony.com>
Git-repo: https://source.codeaurora.org/quic/la/kernel/msm-4.14/commit/?h=msm-4.14
Git-commit: faff2c6cd879b6878ae2998eeb10f570e91dc597
Signed-off-by: Swetha Chikkaboraiah <schikk@codeaurora.org>
Signed-off-by: TogoFire <italomellopereira@gmail.com>