netns: actually test for routing loops
We previously removed the restriction on looping to self, and then added
a test to make sure the kernel didn't blow up during a routing loop. The
kernel didn't blow up, thankfully, but on certain architectures where
skb fragmentation is easier, such as ppc64, the skbs weren't actually
being discarded after a few rounds through. But the test wasn't catching
this. So actually test explicitly for massive increases in tx to see if
we have a routing loop. Note that the actual loop problem will need to
be addressed in a different commit.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Change-Id: I60f48d69560979cfd49b00c653e5e9ec128da478
(cherry picked from commit d1fd106eee14c61b7d9cecaab676bdf87473aa90)
Signed-off-by: TogoFire <togofire@mailfence.com>