fingerprint: goodix: Inline some functions
An inline function is one for which the compiler copies the code from the
function definition directly into the code of the calling function rather
than creating a separate set of instructions in memory. This is mostly done
for small computations, means a function having less codes so that we can
ignore its overhead and perform faster execution.
Since goodix_fb_state_chg_callback() and proximity_state_set() are not that
big and complex, mark them as inline for the compiler to optimize them.

Change-Id: f08db8d746c4c4ce280994c92f5f9fe1f07aedbf
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: TogoFire <togofire@mailfence.com>