powerpc/pseries: fix possible memory leak in ibmebus_bus_init()
[ Upstream commit afda85b963c12947e298ad85d757e333aa40fd74 ]

If device_register() returns error in ibmebus_bus_init(), name of kobject
which is allocated in dev_set_name() called in device_add() is leaked.

As comment of device_add() says, it should call put_device() to drop
the reference count that was set in device_initialize() when it fails,
so the name can be freed in kobject_cleanup().

Signed-off-by: ruanjinjie <ruanjinjie@huawei.com>
Change-Id: I94ba4420aba4991b29dc70c8ffc4c793d74ad9bf
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://msgid.link/20221110011929.3709774-1-ruanjinjie@huawei.com
Signed-off-by: Sasha Levin <sashal@kernel.org>
(cherry picked from commit a4e7691eae30d45041fa153df2f843093ded98af)
Signed-off-by: TogoFire <togofire@mailfence.com>