i40e: fix build warnings in i40e_alloc.h
Not upstream as it was fixed in a much larger api change in newer
kernels.

gcc-13 rightfully complains that enum is not the same as an int, so fix
up the function prototypes in i40e_alloc.h to be correct, solving a
bunch of build warnings.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Change-Id: Ia73c44311f848fa33381d50ce56ad672989606b8
(cherry picked from commit 89765762d2d2a7418738ecda75e8d475d71c3f25)
Signed-off-by: TogoFire <togofire@mailfence.com>