udf: Fix uninitialized array access for some pathnames
[ Upstream commit 028f6055c912588e6f72722d89c30b401bbcf013 ]

For filenames that begin with . and are between 2 and 5 characters long,
UDF charset conversion code would read uninitialized memory in the
output buffer. The only practical impact is that the name may be prepended a
"unification hash" when it is not actually needed but still it is good
to fix this.

Reported-by: syzbot+cd311b1e43cc25f90d18@syzkaller.appspotmail.com
Link: https://lore.kernel.org/all/000000000000e2638a05fe9dc8f9@google.com
Signed-off-by: Jan Kara <jack@suse.cz>
Change-Id: Ic08b7dcf03c8207e360dcc6f75d62772ad119cdc
Signed-off-by: Sasha Levin <sashal@kernel.org>
(cherry picked from commit 856fbe8befbf911a48ef52dd5436d6cf8f2c3c54)
Signed-off-by: TogoFire <togofire@mailfence.com>