iommu: Fix missing return check of arm_lpae_init_pte
UAF scenario may occur in clients with EL1 privileges for
iova mappings when we miss to check the return value of
arm_lpae_init_pte which may lead to an PTE be counted as
it was set even if it was already existing. This can cause a
dangling IOMMU PTE to be left mapped pointing to a
freed object and cause UAF in the client if the dangling PTE
is accessed after a failed unmap operation.

Fixes: 27de1978c331 ("ANDROID: GKI: iommu/io-pgtable-arm: LPAE related updates by vendor")
Signed-off-by: Pratyush Brahma <quic_pbrahma@quicinc.com>
Change-Id: I31a08e17acb39e39818cbfaf21201587cb62f137
(cherry picked from commit 23ef8b4e24d539e6926d537b85eb3f25b219c2a2)
Signed-off-by: TogoFire <togofire@mailfence.com>